+++ changelog.md
@@ -37,3 +37,4 @@
  * Reworked Parallel for the Rotary Hearth Furnace and Bulk Blast Chiller
  * Deprecated Chemical Plant multiblock for implementation in another mod. It will be REMOVED in the following update.
  * Fixed Graphene Foils missing for the Molybdenum Disilicide Coil recipe
+ * Updated Molten fluid generation system to use the new GTCEu MetaFluids API